+++ tests/FreeElephants/JsonApiToolkit/DTO/Example/OneRelationships.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace FreeElephants\JsonApiToolkit\DTO\Example;
+
+use FreeElephants\JsonApiToolkit\DTO\AbstractRelationships;
+use FreeElephants\JsonApiToolkit\DTO\RelationshipToOne;
+
+class OneRelationships extends AbstractRelationships
+{
+    public RelationshipToOne $one;
+}